Completed first implementation of dbench data processing
- Added latency plot for dbench
- Implemented plot_data()

Note that dbench output needs to have continuous timestamps i.e. if multiple
runs are done with multiple processes (e.g. 1, 2, 4) the 6th (seconds)
column resets to 1 with every new run and that makes the resulting plots
really confusing. A helper script for preprocessing the dbench output
would be a good, modular way for dealing with such data files.